refactor(sync): consolidate workspace and Go setup managers
- Replace WorkspaceConfigManager, WorkspaceCheckoutManager, and GoWorkspaceManager
  with unified ConfigManager and WorkspaceManager services
- Update syncCommand to use new WorkspaceManager methods for checkout and Go workspace setup
- Remove deprecated libs: go-workspace-manager.ts, interactive-prompt-manager.ts,
  workspace-checkout-manager.ts, and workspace-config-manager.ts files
- Simplify imports and reduce code duplication for workspace syncing and Go setup logic
